TaskList and CaseOfErrors components part 2
