+++ src/components/CaseOfErrorMessage/CaseOfErrorMessage.tsx
@@ -1,9 +1,8 @@
-/* eslint-disable jsx-a11y/control-has-associated-label */
-
 import { useTodo } from '../../provider/todoProvider';
+import { Button } from './Button';
 
 export const CaseOfErrorMessage = () => {
-  const { error, closeErrorMessage } = useTodo();
+  const { error } = useTodo();
 
   return (
     <div
@@ -12,12 +11,7 @@ export const CaseOfErrorMessage = () => {
         ? 'notification is-danger is-light has-text-weight-normal'
         : 'notification is-danger is-light has-text-weight-normal hidden'}
     >
-      <button
-        data-cy="HideErrorButton"
-        type="button"
-        className="delete"
-        onClick={closeErrorMessage}
-      />
+      <Button />
 
       {error}
     </div>
