+++ src/components/TaskList/Task/EditInput/index.tsx
@@ -1,0 +1,1 @@
+export * from './EditInput';
