+++ src/components/TaskList/Task/Loader/index.tsx
@@ -1,0 +1,1 @@
+export * from './Loader';
