+++ src/components/TaskList/Task/RemoveTaskButton/index.tsx
@@ -1,0 +1,1 @@
+export * from './RemoveTaskButton';
