+++ src/components/TaskList/Task/ToggleInput/ToggleInput.tsx
@@ -1,0 +1,25 @@
+import { useTodo } from '../../../../provider/todoProvider';
+import { Todo } from '../../../../types/Todo';
+
+type Props = {
+  todo: Todo,
+};
+
+export const ToggleInput = ({ todo }: Props) => {
+  const {
+    toggleCompletedTodos,
+  } = useTodo();
+
+  return (
+    <label className="todo__status-label">
+      <input
+        data-cy="TodoStatus"
+        type="checkbox"
+        className="todo__status"
+        onChange={() => toggleCompletedTodos(todo)}
+        checked={!!todo.completed}
+      />
+    </label>
+
+  );
+};
